requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
databases
alembic
pydantic
bcrypt
aiomysql
python-dotenv
starlette
starlette-exporter
gunicorn
python-jose[cryptography]
redis
captcha
pytest
httpx
Faker
isort
black